requirements.txt
# --- Vector DB (app/db/vectordb.py) ---
lib/actiancortex-0.1.0b1-py3-none-any.whl 
numpy>=2.2
grpcio>=1.68
pydantic>=2.10

# --- Tier 1 (Linkup web search) ---
linkup-sdk==0.11.0

# --- Tier 3 (Gemini) ---
google-genai>=1.63.0

# --- Tier 2 (Local LLM: ExecuTorch + Transformers) ---
torch>=2.6.0
transformers>=4.51.0
optimum-executorch>=0.4.0
sentence-transformers>=5.0.0

# --- Common for env vars ---
python-dotenv


# API runtime
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.0
